Extract deep-link/invite parsers out of AppViewModel
Move SessionDeepLinkParser and BuddyInviteTokenParser (with their four
private helpers) out of AppViewModel.swift into dedicated files under
Navigation/. Same-module visibility is preserved, so AppViewModel's thin
wrapper methods (handleIncomingURL, extractBuddyToken) keep calling the
extracted parsers unchanged. No logic changes.

Closes #418
